pyproject.toml
[project]
name = "quotaarchive"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "rozari0", email = "61970262+rozari0@users.noreply.github.com" }
]
dependencies = [
    "django>=5.0.7",
    "django-taggit>=5.0.1",
    "django-crispy-forms>=2.2",
    "crispy-bootstrap5>=2024.2",
    "python-slugify>=8.0.4",
    "django-bleach>=3.1.0",
    "python-decouple>=3.8",
    "dj-database-url>=2.2.0",
    "django-tinymce>=4.1.0",
    "django-autoslug>=1.9.9",
    "crispy-tailwind>=1.0.3",
    "django-admin-interface>=0.28.8",
]
readme = "README.md"
requires-python = ">= 3.8"

[tool.rye]
managed = true
virtual = true
dev-dependencies = []
